requirements.txt
flask
flask-sqlalchemy
flask-user==1.0.2.2
bs4
requests
scikit-learn
pandas
